+++ Source/AdminApp/SystemClasses/SystemSingleton.cs
@@ -25,7 +25,7 @@ namespace AdminApp.SystemClasses
             }
             public static string Login;
             public static Guid ID;
-            public static int TelegramID;
+            public static long TelegramID;
             public static string FirstName;
             public static string LastName;
             public static string FullName;
